refactor(active_inference): extract figure_provenance to satisfy line-count budget
The integrity hardening pushed integration_audit.py to 1008 lines, over the
950-line composability gate (module_line_count_check.py). Extract the cohesive
figure source-provenance unit (build_figure_source_map + the existence-aware
_figure_sources_mapped/_is_deferred_source/_source_path_exists helpers) into a
new roadmap_tracks/figure_provenance.py and re-import. integration_audit.py ->
919 lines (gate passes); figure_provenance.py 104. No behavior change: build is
identical (all_figures_mapped True, 20 figures), roadmap suite 10/10, ruff/mypy
clean, check-all-exports 0 violations.
